[FEAT] #75 base_route_id 받아오기 (location 적용)

+++ app/src/main/java/com/timi/seulseul/data/model/Location.kt
@@ -2,7 +2,7 @@ package com.timi.seulseul.data.model
 
 // Request Body
 data class Location(
-    val id: Int,
+    val id: Long,
     val startX: String,
     val startY: String,
     val dayInfo: String
@@ -10,7 +10,7 @@ data class Location(
 
 // PATCH Request Body
 data class PatchLocation(
-    val id: Int,
+    val id: Long,
     val startX: String,
     val startY: String
 )
@@ -23,7 +23,7 @@ data class LocationResponse(
 
 // Response - data 필드 내부 구조
 data class LocationData(
-    val id: Int,
+    val id: Long,
     val startX: String,
     val startY: String,
     val dayInfo: String
